requirements.txt
git+https://github.com/PelvisVR/deepdrr.git@direct_voxelize
pyzmq
pycapnp
typer
scipy
joblib
more_itertools